+++ src/entries/components/account/ReturnComparison/index.js
@@ -1,0 +1,1 @@
+export { default } from './ReturnComparison';
