--- src/entries/components/returnComparison/index.js
@@ -1,3 +1,0 @@
-export { default } from './ReturnComparison';
-export { default as ReturnComparisonDateSelect } from './ReturnComparisonDateSelect';
-export * from './api';
